Patreon pitch and reason options
Added a text field in the Patreon options for a pitch page ulr entry and
pitch page reason.  The url is the full url to a page that will be used
to pitch people your Patreon campaign.  In addition, the reason section
is whatever text you want the viewer to see at the top of the requested
page as a reason they are not seeing the content.  in the reason string,
if you use "$patreon_level" it will show the requested pages patreon
level.

Example "you must be at least a $patreon_level patron to view this
content"