Show pitch page on protected content
This will fetch the patreon settings of the pitch page url and reason.
The pitch page will be displayed as the content if a viewer is trying to
see patreon protected content and not that level of subscribership.  It
will also at the top show them the reason they cannot view the content
by showing whatever the pitch reason says in the patreon options.