Fix reconnect flow for deleted oauth clients (#178)
### Problem
The reconnect flow was not working properly when attempting to reconnect
if the
client had been deleted on patreon. When completing the reconnect flow,
the
plugin code verified that the `client_id` had not changed. This is an
incorrect
assumption. I think there was an attempt to fix this
[here](https://github.com/Patreon/patreon-wordpress/pull/153/files#diff-e1b26a9c72951cd70b42f44a9ba780b9bde831fda64620d651f631d5e0c158eeR361),
but the check might've
been added to the incorrect if block.


### Solution
- Do not enforce a check that the client id didn't change during
reconnect
- While I'm here, merge the reconnect & connect flow logic since they
were basically the
same
- Extract the `patreon-authorization` logic in a separate method to make
the root routing
method more readable