Return object during creator token refresh only if the tokens were refreshed (#185)
### Problem
The callers of `refresh_cretor_access_token` are looking for a truthy value
to decide whether the refresh was successful. A recent change modified 
`refresh_token` method and it now always returns a truthy value (it
includes http response code). This means that even 401 and 429
calls are considered as successful token refresh.

### Solution
Return token refresh data only if it included access & refresh tokens.