Set correct variable for refresh token expiration (#188)
### Problem
The token expiration was incorrectly set (PR
https://github.com/Patreon/patreon-wordpress/pull/186) on the scope
option and not the
expiration one. This meant that the expiration was considered being
empty forcing
the plugin to perform creator token refresh.

### Solution
Set the `patreon-creators-refresh-token-expiration` option with the new
expiration.
The `patreon-creators-access-token-scope` is only set & never used, the
incorrect
value won't cause any issues. The tokens and thus scope info should
rotate within 30
days.