Add a queryparam to token refresh to indicate if refresh is for creator access token (#192)
### Problem
`POST /api/oauth2/token` results with high number of HTTP 401s. It's not
clear what's causing this. I would like to get more data on the api side
to understand what's triggering a token refresh.

### Solution
Add a `is_creator_access` queryparam to the token refresh call. This
queryparam is set based in the `$disable_app_on_auth_err` value, which
is usually set to `true` for creator access tokens.

This is temporary and can be eventually cleaned up.